+++ src/lib.rs
@@ -304,16 +304,14 @@ pub use types::dynamic::{AccessMode, DynamicReadLock, DynamicWriteLock, Variant}
 pub use types::FloatWrapper;
 
 #[cfg(feature = "internals")]
-pub use tokenizer::{get_next_token, parse_string_literal};
+pub use types::StringsInterner;
 
 #[cfg(feature = "internals")]
 pub use tokenizer::{
-    is_valid_function_name, is_valid_identifier, InputStream, MultiInputsStream, Span, Token,
-    TokenIterator, TokenizeState, TokenizerControl, TokenizerControlBlock,
+    get_next_token, is_valid_function_name, is_valid_identifier, parse_string_literal, InputStream,
+    MultiInputsStream, Span, Token, TokenIterator, TokenizeState, TokenizerControl,
+    TokenizerControlBlock,
 };
-
-#[cfg(feature = "internals")]
-pub use types::StringsInterner;
 
 #[cfg(feature = "internals")]
 pub use parser::ParseState;
@@ -339,6 +337,10 @@ pub use ast::EncapsulatedEnviron;
 
 #[cfg(feature = "internals")]
 pub use eval::{Caches, FnResolutionCache, FnResolutionCacheEntry, GlobalRuntimeState};
+
+#[cfg(feature = "internals")]
+#[allow(deprecated)]
+pub use func::NativeCallContextStore;
 
 #[cfg(feature = "internals")]
 #[cfg(feature = "metadata")]